refactor: move and rename types file to detail.ts (#78)
- QnaDetail 관련 타입 파일명 변경 및 경로 이동 (types.ts -> detail.ts)
- 변경된 경로 참조하는 관련파잉ㄹ 일괄 수정
