+++ src/features/qna-detail/components/AnswerCard.tsx
@@ -3,9 +3,8 @@ import { MessageCircle } from 'lucide-react'
 import { Avatar, ModalButton } from '@/components'
 import { useCommentSort } from '@/hooks'
 import type { SortType } from '@/hooks/useCommentSort'
+import type { QnaAnswer } from '@/types'
 import { cn, formatTimeAgo } from '@/utils'
-
-import type { QnaAnswer } from '../types'
 
 type AnswerCardProps = {
   answer: QnaAnswer
