+++ src/types/api-response/detail.ts
