+++ src/types/index.ts
@@ -1,3 +1,4 @@
 // types
 export * from './api-response/category'
+export * from './api-response/detail'
 export * from './api-response/qna'
